Added first spec for the #phrase method
Normalised the handle for stripping out the @ so that we can assign and use the
handle in the message easily.

The previous change (checking handle existence) resulted in having to re-run the
vcr requests because of the extra request so the features needed to change
slightly to reflect the new scores.